requirements.txt

Django==4.2.11
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-filter==24.1
django-simple-history==3.5.0
django-phonenumber-field==7.3.0
django-versatileimagefield==3.1
django-dirtyfields==1.9.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
icecream==2.1.3
google-api-python-client-stubs==1.25.0

